requirements.txt
dj-database-url==0.5.0
Django==2.0.2
django-cors-headers==2.4.0
django-extensions==2.1.0
django-filter==2.0.0
django-location-field==2.0.3
djangorestframework==3.8.2
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
whitenoise==3.3.1
